refactor: extract useChat hook and separate ChatInputBox and ChatMessageBubble components
